Type coneversion fix
git-svn-id: https://svn.physiomeproject.org/svn/cmiss/cmgui/trunk@1678 3c1deb5b-d424-0410-962d-aba41a686d42

+++ source/unemap/page_window.c
@@ -2392,7 +2392,7 @@ named file.
 							temp_file_name);
 						save_write_signal_file_background_data->page_window=page_window;
 						if (unemap_get_samples_acquired_background(0,
-							page_window->number_of_samples_to_save,
+							(int)page_window->number_of_samples_to_save,
 							save_write_signal_file_background,
 							(void *)save_write_signal_file_background_data))
 						{
